+++ ExperPort/Analysis/ProAnti/pa_analyze.m
@@ -104,7 +104,7 @@ handles = [];
 colors  = [0 0 1 ; 1 0 1 ; 1 0 0 ; 0 1 0]; 
 
 if isempty(fignum),  %#ok<NODEF>
-  figure; fignum = gcf;
+  figure; fignum = double(gcf);
 end;
 if ~ishandle(fignum), 
   figure(fignum);
